Passcode input working at it's own component
